fix Util#createFetchableDictionary

+++ src/main/java/com/skyblock/skyblock/features/potions/effects/SpeedEffect.java
@@ -9,7 +9,7 @@ import java.util.function.Function;
 
 public class SpeedEffect extends PotionEffect {
 
-    private static final Function<Integer, Double> calculateSpeedBonus = (level) -> Util.createFetchableDictionary(level - 1, 5, 10, 15, 20, 25, 30, 35, 40);
+    private static final Function<Integer, Double> calculateSpeedBonus = (level) -> (double) Util.createFetchableDictionary(level - 1, 5, 10, 15, 20, 25, 30, 35, 40);
 
     public SpeedEffect(SkyblockPlayer player, String ignoredName, int amplifier, double duration, boolean alreadyStarted) {
         super(player, "Speed", amplifier, duration, alreadyStarted, "&7Grants &a+" + calculateSpeedBonus.apply(amplifier) + " &f" + SkyblockStat.SPEED.getIcon() + " Speed&7.");
